requirements.txt
requests
beautifulsoup4
google-cloud-firestore
google-cloud-storage
google-cloud-speech
google-cloud-pubsub
flask
gunicorn